CargaMasivaPlanMedios/requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-bigquery
google-cloud-core
pandas
pyarrow
datetime
pandas-gbq
openpyxl